+++ test/unit/fuzz-tools.test.js
@@ -265,6 +265,7 @@ import * as proteinsV274 from '../../lib/proteins-v274.js';
 import * as fibrosisV275 from '../../lib/fibrosis-v275.js';
 import * as nutritionV276 from '../../lib/nutrition-v276.js';
 import * as renalV277 from '../../lib/renal-v277.js';
+import * as pedsSepsisV278 from '../../lib/peds-sepsis-v278.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -474,6 +475,7 @@ const MODULES = {
   'fibrosis-v275.js': fibrosisV275,
   'nutrition-v276.js': nutritionV276,
   'renal-v277.js': renalV277,
+  'peds-sepsis-v278.js': pedsSepsisV278,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];